[breadboard-cli] Clean up the debug module
